% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{close_mtbl}
\alias{close_mtbl}
\title{Close an mtbl file}
\usage{
close_mtbl(x)
}
\arguments{
\item{x}{an open mtbl file handle}
}
\description{
Will be really useful if we ever implement the write components of the API.
}

