% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mtbl_export_values}
\alias{mtbl_export_values}
\title{Export mtbl \emph{values} to a file}
\usage{
mtbl_export_values(x, path)
}
\arguments{
\item{x}{an mtbl file opened with \code{read_mtbl()}}

\item{path}{full path to output file}
}
\description{
Given an \code{mtbl} file opened with \code{read_mtbl()}, this function
will export each value from the mtbl key/value pair to \code{path} with
each value record terminated by a newline.
}
\details{
The use-case driving this function is exporting the value side of
Project Sonar mtbl files since it would help in processing \code{jq}
pipelines.
}
\examples{
mtbl <- read_mtbl(system.file("extdata/sample.mtbl", package="mtblr"))
mtbl_export(mtbl, tempfile())
}

