% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mtbl_get}
\alias{mtbl_get}
\title{Retrieve a key/value pair from an mtbl by key}
\usage{
mtbl_get(x, key)
}
\arguments{
\item{x}{an mtbl file opened with \code{read_mtbl()}}

\item{key}{string (exact) to look for in the keys}
}
\description{
In theory there are no duplicates, but this still
returs an iterator that you'll need to use \code{mtbl_next()} with.
}
\examples{
mtbl <- read_mtbl(system.file("extdata/sample.mtbl", package="mtblr"))
mtbl_next(mtbl_get(mtbl, "aa"))
}

