% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mtbl_get_prefix}
\alias{mtbl_get_prefix}
\title{Retrieve value(s) for a given key in an mtbl}
\usage{
mtbl_get_prefix(x, key)
}
\arguments{
\item{x}{an mtbl file opened with \code{read_mtbl()}}

\item{key}{string (prefix) to look for in the keys}
}
\description{
Retrieve value(s) for a given key in an mtbl
}
\examples{
mtbl <- read_mtbl(system.file("extdata/sample.mtbl", package="mtblr"))
mtbl_next(mtbl_get_prefix(mtbl, "a"))
}

