% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mtbl_keys}
\alias{mtbl_keys}
\title{Return all the keys of an mtbl}
\usage{
mtbl_keys(x)
}
\description{
Return all the keys of an mtbl
}
\examples{
mtbl <- read_mtbl(system.file("extdata/sample.mtbl", package="mtblr"))
head(mtbl_keys(mtbl))
}

