% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mtbl_length}
\alias{mtbl_length}
\title{Retrieve the number of records in an open mtbl file}
\usage{
mtbl_length(x)
}
\arguments{
\item{x}{mtbl file opened with \code{read_mtbl()}}
}
\description{
Retrieve the number of records in an open mtbl file
}
\examples{
mtbl <- read_mtbl(system.file("extdata/sample.mtbl", package="mtblr"))
mtbl_length(mtbl)
}

