% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mtbl_next}
\alias{mtbl_next}
\title{Iterate over mtbl file entries}
\usage{
mtbl_next(x, n = 100L)
}
\arguments{
\item{x}{an iterator started with \code{mtbl_map()} or \code{mtbl_get()}}

\item{n}{number of records to return for this iteration. It's inefficient to
sequentially iterate over large mtbl file but large values of \code{n}
are also inefficent. \code{100} has been an optimal value in mtbl
files used in Project Sonar. YMMV.}
}
\description{
Works with \code{mtbl_map()} and \code{mtbl_get()}
}
\details{
If the number of available entries is less than \code{n}, a \code{list}
of the actual number of remaining entries is returned. Otherwise a \code{list}
of \code{n} entries is returned.
}
\examples{
mtbl <- read_mtbl(system.file("extdata/sample.mtbl", package="mtblr"))
mp <- mtbl_map(mtbl)
(mtbl_next(mp))
mtbl_iter_close(mp)
}

