% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mtbl_values}
\alias{mtbl_values}
\title{Return all the values of an mtbl}
\usage{
mtbl_values(x)
}
\arguments{
\item{x}{an mtbl file opened with \code{read_mtbl()}}
}
\description{
Return all the values of an mtbl
}
\examples{
mtbl <- read_mtbl(system.file("extdata/sample.mtbl", package="mtblr"))
head(mtbl_values(mtbl))
}

