% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{read_mtbl}
\alias{read_mtbl}
\title{Open an mtbl file}
\usage{
read_mtbl(path)
}
\arguments{
\item{path}{full path to mtbl file}
}
\description{
Open an mtbl file
}
\examples{
mtbl <- read_mtbl(system.file("extdata/sample.mtbl", package="mtblr"))
}

